WIP migration back to nvim, but without nvchad.
nvchad is too opinionated. Finding myself wanting to undo or change things
that are hard, and the trade-off of that against the benefits isn't
worth it.

And I use vim bindings for too many things to not use it as my default
editor. Going to start slowly migrating back to vim and re-building up
my own config.